Added terrain, player and 1st animation
Added a prototype of the terrain.
Added the player scripts, menus and hp adapted from lvl 1.
Created the colliders of the map, the shadows and the 3D terrain
heights.
Implemented an in-game animation that shows a Demon appearing from the
volcan.
Imported some assets from the lvl2 demo showed at class.
Added scripts for new npcs and some other small things.